8913 Move resolution summary to newly created CompleteResolutionSummary component
