+++ src/components/CreateResolution/index.ts
@@ -1,5 +1,7 @@
 import CompleteResolution from './CompleteResolution.vue'
+import CompleteResolutionSummary from './CompleteResolutionSummary.vue'
 
 export {
-  CompleteResolution
+  CompleteResolution,
+  CompleteResolutionSummary
 }
